Propagate dependencies into AF output args
Per issue #67, the e:dependency template in adms.implicit.xml will
propagate probe and variable dependencies from analog function
arguments into the top-level expression in which the function is
called (generally the RHS of an assignment), but does NOT do the same
thing for any output arguments that might exist in the analog
function.

Some new CMC standard models (e.g. the MVSG-HV model) use analog
functions in this manner, and it would be good to fix it.

This commit augments the e:dependency template so that it completely
propagates all dependencies from function arguments into any output
variables that might exist.  It is a no-op if there are no such
function output arguments.

This commit also obsoletes pull request #68.
Unlike that PR, this one makes sure to perform dependency checking on
the input arguments before copying their probe dependencies, and
unlike that PR, it also copies their variable dependencies.  Also
unlike that code, it will correctly handle output arguments that are
not at the beginning of the argument list, because it selects all
arguments in the function definition in order and uses their position
in the global list rather than selecting only the output arguments and
using their position in the truncated list.

Issue #67
PR #68