Move ADMS include directory to back of search list
The previous augmentation of the include path resulted in the following
search order.

    1. ADMS include directory.
    2. Current working directory.
    3. User specified include directories.

This means that the ADMS standard header files, namely constants.vams
and disciplines.vams, would always take precedence over user supplied
files of the same name. This commit changes the search order to the
following.

    1. Current working directory.
    2. User specified include directories.
    3. ADMS include directory.

This way users are not forced to delete or edit the incomplete header
files in the ADMS include directory.