Allow semicolon after discipline/nature identifier
Version 2.3.0 of the Verilog-AMS Language Reference Manual introduced
an optional semicolon after a nature and discipline identifier in a
nature and discipline declaration, respectively; e.g.,

    nature identifier;
        ...
    endnature

and

    discipline identifier;
        ...
    enddiscipline

are now also legal syntax.

ADMS behaves the same as when the semicolon is absent.

This brings ADMS one step closer to being able to parse the most
recent Accellera standard disciplines header files.

This commit fixes GitHub issue #39.