Fix bad grammar rule for optional semicolon
My previous attempt to allow for an optional semicolon after a nature
and discipline identifier in a nature and discipline declaration,
respectively, resulted in the semicolon now being mandatory. This
commit actually makes it optional.

I have tested this with Accellera's v2.2.0 and v2.3.1 standard headers as
well as ADMS's v2.3.7 standard header.