Move renamed functions into libadmsElement
The adms_dirname function needs to be used in preprocessorLex.l, but
it is currently a static function declared in admsXml.c. Making it
accessible from libadmsElement allows both files to use it.

The dependencies, adms_split_new and adms_free_strlist, have also
been relocated, as well as the related functions adms_basename and
adms_filename, for consistency.