Flag LRM-compliant output variables in adms data tree
Per issue Qucs/ADMS#72, the implicit rules in adms.implicit.xml only
flag a variable in the data tree with "output=yes" in the data tree if
they have the attribute 'ask="yes"'.

The LRM (Verilog-AMS LRM version 2.4, section 3.2.1) says that
module-scoped variables with either "desc" or "units" attributes
should be designated as output variables.

This commit modifies adms.implicit.xml to set the "output" field of the
data tree to "yes" if the variable is not an input
variable (parameter) and is module scoped, and if it has either a "desc"
or a "units" (or both) attribute.

The Xyce team has been using a patched adms.implicit.xml with this
modification for a couple of years, and I had simply forgotten that
I'd opened issue Qucs/ADMS#72 asking that it be done in ADMS itself.

Qucs/ADMS#72